Logic added to progress through the quiz
Buttons change colors on hover. On click they evaluate if the answer is correct and if so add a point to the score. Changes color of button to reflect if correct. Next Question button then appears. Cycle continues until last question is reached (hard coded for four questions right now). Button then changes to 'Check Score.'